refactor filenames based on discussion, propose `regrid.regrid_dataset()` as refactoring of `regrid.regrid()`
