+++ xscen/aggregate.py
@@ -11,7 +11,7 @@ import xesmf as xe
 from shapely.geometry import Polygon
 
 from .config import parse_config
-from .extraction import clisops_subset
+from .extract import clisops_subset
 
 logger = logging.getLogger(__name__)
